improved the convenience of working with colors and dropper, now everything will be more convenient. For now, in the experimental
Now when you click on a "T" in Sketch or in Inpaint Sketch, an open color selection panel appears next to the mouse, and it works by incrementing.

You may or may not have problems if you select color with a very large image width greater than 1200px and a strong zoom.