Added the ability to zoom in with the + and - keys  (#20)
* adding context menu support

I plan on making this work for more things in the future

* added more options and added some style

* working on right click settings changes

* fix

* Remove old code

* Mini-Fix to make the context menu only on the image area

* Fix some bugs and improve open brush colors hotkey

1) edited comments
2)Made a contextual me with a large index, so that it would overlap the picture
3) Made the context menu close in 200ms after you moved the mouse away from it.
3) Changed menu Settings to Change hotkeys so it would be understandable and made it so the user would not fall in modal window
4) Changed a little Open Brush Color, now when you click on the color selection panel opens, I think it is more convenient
5) Fixed bug with duplicate overlap and open brush color

Translated with www.DeepL.com/Translator (free version)

* Update README.MD

Added a description, for the RU language

* The solution to solve problem #1 in ReadmeMD.

When the user uploads a file we reset the zoom and the picture is always loaded correctly without the zoom.

* Update README.MD

* Update README.MD

* Update README.MD

* Update README.MD

* Fixed a bug with the color panel

There was a bug where if you switched to sketch or inpaint sketch, the panel would crash and not display correctly.
I fixed that bug.

* Overlap triggered during image preview

During the image preview, right-clicking on any area will trigger an Overlap

Thanks XpusT for the idea, in his video it was a bug, now it's a feature.

* Update README.MD

* Update canvas-zoom.js

* Update canvas-zoom.js

* Update README.MD

* Update README.MD

* Update canvas-zoom.js

* Now hotkeys work in any layout, changed the name of the variable localStorage to avoid errors when updating

I will add figures later, the name of the variable changed because now hotkeys look different and if someone has an old entry in the localStorage, it is easier to make another than, do checks

* Update README.MD

* Fixed duplicate hotkeys

* Added a new move mode, with a single button, the mode can be changed in the context menu.

Added a new move mode, as in Blender, now when you click on G (you can configure), the picture moves behind the mouse, that undo the move you need to click on G again.

My code can definitely be improved

* Made the processing of most actions only inside the image area, fixed bugs with localStorage

After testing it, I realized that it's better when you hold down the key and drag than when you press it twice.

Fixed bugs with localStorage, there were errors if there was no keys in memory. Took out a little bit of logic.

* Change readme

* Update canvas-zoom.js

* Update README.MD

* Prevented disabling Ctr-Z and any combination of Ctr + "Key" on text2img

* Update README.MD

* improved the convenience of working with colors and dropper, now everything will be more convenient. For now, in the experimental

Now when you click on a "T" in Sketch or in Inpaint Sketch, an open color selection panel appears next to the mouse, and it works by incrementing.

You may or may not have problems if you select color with a very large image width greater than 1200px and a strong zoom.

* Update README.MD

* Update canvas-zoom.js

* turning this into a feature

* Add a description of the feature

* Now Ctr-Z works only in the image area, it also does not cancel actions in other tabs

* Update README.MD

* I fixed the color panel binding to the mouse, now it works without any problems

Now the appearance of the color bar next to the mouse will always be in the right place. It is necessary that the mouse must be in the area of the image

* Update README.MD

* Update README.MD

* Reset height and width if user stretched the area.

Little code optimization

* Removed unused code, deleted unnecessary comments, refactored some parts of the code, added comments for functions

Removed the modal window code because it is not used in the code. Refactored the toggleBrushPanel function, added comments to some parts of the code to make it clearer

* Fixed the context menu to show up

I made a fix for the context menu to appear with a random click, now it disappears after 0.5 seconds if the user does not hover the mouse over it

* minor changes

* Fix minnor bug

* Update README.MD

* Update README.MD

* Update README.MD

* Update README.MD

* Added the ability to zoom in with the shift + and shift - keys

For users who do not have a mouse or for those who do not want to use the wheel added the ability to zoom through the buttons, numpad also works

* Update README.MD

* Disabled Overlap during previews

I decided to disable the overlap that obscures the preview, I think it does more harm than good.

* Update README.MD

---------

Co-authored-by: richrobber2 <richardrgguno@hotmail.com>