Destructuring: Instead of accessing e.deltaY multiple times, we can destructure it directly in the function parameters. This makes the code cleaner and easier to read.
Early Return: Instead of wrapping the adjustBrushSize call in an if statement, we can return early if the condition is not met. This reduces the level of indentation and makes the code more readable.

Implicit Return: In arrow functions, if the function body consists of a single statement, you can omit the braces and the return keyword. This makes the code more concise. However, in this case, we need to explicitly return false to prevent the default action.